Move task_storage code under the bpf tag
